+++ src/components/new-image/_tests_/new-image-test-cases.js
@@ -1,0 +1,13 @@
+import NewImage from '../new-image';
+
+const testCases = {};
+
+testCases.basic = {
+  component: NewImage,
+  description: 'Basic',
+  props: {
+    size: '60'
+  }
+};
+
+export { testCases };