+++ src/components/note/note.js
@@ -33,19 +33,7 @@ class Note extends React.Component {
     if (props.title) {
       titleText = props.title;
     }
-
-    const createStyles = (base, theme) => {
-      let styles = {};
-      for (let style in base) {
-        styles[style] = base[style];
-      }
-      for (let style in theme) {
-        styles[style] = theme[style];
-      }
-      return styles;
-    };
-
-    const themeNote = createStyles(
+    const themeNote = Object.assign(
       themes['base'],
       themes[props.theme] || themes['default']
     );