+++ src/helpers/highlight-json.js
@@ -1,0 +1,8 @@
+import Prism from 'prismjs';
+import 'prismjs/components/prism-json';
+
+function highlightJson(rawCode) {
+  return rawCode ? Prism.highlight(rawCode, Prism.languages['json']) : '';
+}
+
+export { highlightJson };