+++ src/components/book-image/book-image.js
@@ -5,32 +5,24 @@ export default class BookImage extends React.Component {
   render() {
     const { props } = this;
     return (
-      <svg
-        id="Layer_1"
-        data-name="Layer 1"
-        xmlns="http://www.w3.org/2000/svg"
-        width={props.size}
-        height={props.size}
-        viewBox="0 0 108 108"
-      >
-        <title>help</title>
+      <svg width={props.size} height={props.size} viewBox="0 0 108 108">
         <circle cx="53" cy="55" r="45" fill="#47b9f3" />
         <rect
-          x="27.2732"
-          y="41.2042"
-          width="52.2541"
-          height="34.5008"
-          rx="5.2192"
-          ry="5.2192"
-          transform="translate(0.6201 -0.5605) rotate(0.6049)"
+          x="27.273"
+          y="41.204"
+          width="52.254"
+          height="34.501"
+          rx="5.219"
+          ry="5.219"
+          transform="rotate(.605 53.4 58.455)"
           fill="#1274c6"
         />
         <path
-          d="M64.1,34.8478s-9.9116-.76-10.0085,8.42-.0922,24.7363-.0922,24.7363,4.823-3.0455,8.79-3.0037l15.4041.1626A1.7881,1.7881,0,0,0,80,63.3935L80.281,36.792A1.7732,1.7732,0,0,0,78.5267,35Z"
+          d="M64.1 34.848s-9.912-.76-10.008 8.42-.093 24.736-.093 24.736S58.822 64.96 62.79 65l15.404.163A1.788 1.788 0 0 0 80 63.393l.281-26.601A1.773 1.773 0 0 0 78.527 35z"
           fill="#309ef9"
         />
         <path
-          d="M43.971,34.157s10.21-.5625,10.1107,8.821S53.997,68,53.997,68s-5.25-2.9574-9.3313-3.0005L28.78,64.8314A1.7863,1.7863,0,0,1,27,63.04l.2881-27.2857A1.7863,1.7863,0,0,1,29.1058,34Z"
+          d="M43.971 34.157s10.21-.563 10.11 8.821S53.998 68 53.998 68s-5.25-2.957-9.331-3L28.78 64.83A1.786 1.786 0 0 1 27 63.04l.288-27.286A1.786 1.786 0 0 1 29.106 34z"
           fill="#0d84fd"
         />
       </svg>
@@ -39,5 +31,9 @@ export default class BookImage extends React.Component {
 }
 
 BookImage.propTypes = {
-  size: PropTypes.string.isRequired
+  size: PropTypes.number
+};
+
+BookImage.defaultProps = {
+  size: 60
 };