+++ src/components/troubleshoot-image/__tests__/troubleshoot-image.test.js
@@ -20,4 +20,22 @@ describe('troubleshoot-image', () => {
       expect(tree).toMatchSnapshot();
     });
   });
+
+  describe(testCases.bigger.description, () => {
+    let testCase;
+    let wrapper;
+    let tree;
+
+    beforeEach(() => {
+      testCase = testCases.bigger;
+      wrapper = renderer.create(
+        React.createElement(testCase.component, testCase.props)
+      );
+      tree = wrapper.toJSON();
+    });
+
+    test('renders as expected', () => {
+      expect(tree).toMatchSnapshot();
+    });
+  });
 });