+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Master
+* Add a new `Note` style with a green background that will be used to show off new or updated products or features. [#162](https://github.com/mapbox/dr-ui/pull/162)
+* Fix `window.scroll()` on `BackToTopButton`. [#167](https://github.com/mapbox/dr-ui/pull/167)
+* Enable babel polyfill on `Search` component. [#165](https://github.com/mapbox/dr-ui/pull/165)
+
 ## 0.19.3
 
 * Fix `Note` in IE11 by replacing Object.assign with a function.[#159](https://github.com/mapbox/dr-ui/pull/159)