+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix `window.scroll()` on `BackToTopButton`. [#167](https://github.com/mapbox/dr-ui/pull/167)
 * Enable babel polyfill on `Search` component. [#165](https://github.com/mapbox/dr-ui/pull/165)
 * Add babel config to helpers and plugins, add eslint plugins to catch new JavaScript features. [#164](https://github.com/mapbox/dr-ui/pull/164)
+* Add `GLWrapper` and `DemoIframe` components. [#166](https://github.com/mapbox/dr-ui/pull/166)
 
 ## 0.19.3