+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 * Enable babel polyfill on `Search` component. [#165](https://github.com/mapbox/dr-ui/pull/165)
 * Add babel config to helpers and plugins, add eslint plugins to catch new JavaScript features. [#164](https://github.com/mapbox/dr-ui/pull/164)
 * Add `GLWrapper` and `DemoIframe` components. [#166](https://github.com/mapbox/dr-ui/pull/166)
+* Refactor image components to fix IE11 bug and set default size. [#171](https://github.com/mapbox/dr-ui/pull/171)
 
 ## 0.19.3
 
-* Fix `Note` in IE11 by replacing Object.assign with a function.[#159](https://github.com/mapbox/dr-ui/pull/159)
+* Fix `Note` in IE11 by replacing Object.assign with a function. [#159](https://github.com/mapbox/dr-ui/pull/159)
 
 ## 0.19.2